refactor: move educator api tests out of api route folder
