+++ src/__tests__/pages/api/educator/saveUnit/[programmeSlug]/[unitSlug].test.ts
@@ -1,8 +1,7 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import { createMocks } from "node-mocks-http";
 
-import handler from "./[programmeSlug]";
-
+import handler from "@/pages/api/educator/saveUnit/[programmeSlug]/[unitSlug]";
 import {
   installMockClerkClient,
   mockServerUser,
@@ -15,11 +14,16 @@ jest.mock("@clerk/nextjs/server");
 
 const updateUserMetadata = jest.fn();
 const getUser = jest.fn();
-const mockGetUserContent = jest.fn();
+
+const mockCreateUserListContent = jest.fn();
+const mockGetUser = jest.fn();
+const mockCreateUser = jest.fn();
 
 jest.mock("@/node-lib/educator-api", () => ({
   getAuthenticatedEducatorApi: jest.fn().mockResolvedValue({
-    getUserContent: () => mockGetUserContent(),
+    createUserListContent: () => mockCreateUserListContent(),
+    getUser: () => mockGetUser(),
+    createUser: () => mockCreateUser(),
   }),
 }));
 
@@ -32,7 +36,7 @@ jest.mock("@/common-lib/error-reporter", () => ({
       mockErrorReporter(...args),
 }));
 
-describe("/api/educator-api/getSavedUnits/[programmeSlug]", () => {
+describe("/api/educator-api/saveUnit/[programmeSlug]/[unitSlug]", () => {
   installMockClerkClient({
     updateUserMetadata,
     getUser,
@@ -41,34 +45,42 @@ describe("/api/educator-api/getSavedUnits/[programmeSlug]", () => {
 
   beforeEach(() => {
     setGetAuth(mockGetAuthSignedIn);
-    mockGetUserContent.mockResolvedValue({
-      users_content: [
-        { users_content_lists: { content: { unit_slug: "unit1" } } },
-        { users_content_lists: { content: { unit_slug: "unit2" } } },
-      ],
+    mockCreateUserListContent.mockResolvedValue({});
+    mockGetUser.mockResolvedValue({
+      user: [{ id: "test-user-id" }],
     });
   });
 
-  it("should return 200 with valid unit slugs for a signed in user", async () => {
+  it("should return 200 for a signed in user", async () => {
     const { req, res } = createMocks<NextApiRequest, NextApiResponse>({
-      method: "GET",
-      query: { programmeSlug: "test-programme" },
+      method: "POST",
+      query: { programmeSlug: "test-programme", unitSlug: "test-unit" },
     });
 
     await handler(req, res);
     expect(res._getStatusCode()).toBe(200);
-    expect(res._getJSONData()).toEqual(["unit1", "unit2"]);
+    expect(mockCreateUser).not.toHaveBeenCalled();
   });
-  it("should return 401 with empty array for a signed out user", async () => {
+  it("should create a user if it does not exist", async () => {
+    mockGetUser.mockResolvedValue({ user: [] });
+
+    const { req, res } = createMocks<NextApiRequest, NextApiResponse>({
+      method: "POST",
+      query: { programmeSlug: "test-programme", unitSlug: "test-unit" },
+    });
+
+    await handler(req, res);
+    expect(mockCreateUser).toHaveBeenCalled();
+  });
+  it("should return 401 for a signed out user", async () => {
     setGetAuth(mockGetAuthSignedOut);
     const { req, res } = createMocks<NextApiRequest, NextApiResponse>({
       method: "GET",
-      query: { programmeSlug: "test-programme" },
+      query: { programmeSlug: "test-programme", unitSlug: "test-unit" },
     });
 
     await handler(req, res);
     expect(res._getStatusCode()).toBe(401);
-    expect(res._getJSONData()).toEqual([]);
   });
   it("should return 400 if programmeSlug is missing or invalid", async () => {
     const { req, res } = createMocks<NextApiRequest, NextApiResponse>({
@@ -80,12 +92,14 @@ describe("/api/educator-api/getSavedUnits/[programmeSlug]", () => {
     expect(res._getStatusCode()).toBe(400);
     expect(res._getData()).toBe("Bad request");
   });
-  it("should report an error", async () => {
-    mockGetUserContent.mockRejectedValue(new Error("Test error"));
+  it("should report an error if the request fails", async () => {
     const { req, res } = createMocks<NextApiRequest, NextApiResponse>({
-      method: "GET",
-      query: { programmeSlug: "test-programme" },
+      method: "POST",
+      query: { programmeSlug: "test-programme", unitSlug: "test-unit" },
     });
+
+    const error = new Error("Test error");
+    mockCreateUserListContent.mockRejectedValue(error);
 
     await handler(req, res);
     expect(res._getStatusCode()).toBe(500);
