+++ src/pages/api/educator/getSavedContentLists/index.ts
